Added a new plugin for the AI hole finder with lattice extension.
Needed to make changes to the base plugin to allow for the lattice extension to be added to the AI hole finder.
More parameters from the grid object can now be passed to the plugins.
These addition will not break the other finders as the parameters are optional.

Added a new protocol to make use of the new plugin.

Changed default protocol to remove the Ptolemy-based protocol.